requirements.txt
# 店舗情報スクレイピングツール 依存関係
# インストール: pip install -r requirements.txt

# Playwright（CLI版用・ブラウザ自動化）
playwright>=1.40.0

# requests + BeautifulSoup（GUI版用）
requests>=2.31.0
beautifulsoup4>=4.12.0

# Gemini API（メインSDK）
google-genai>=1.0.0

# Gemini API（フォールバック用旧SDK）
google-generativeai>=0.8.0

# OpenAI SDK（Perplexity API用）
openai>=1.0.0

# 環境変数読み込み
python-dotenv>=1.0.0

# リッチな出力表示
rich>=13.0.0

# GUI
streamlit>=1.30.0
pandas>=2.0.0

# Excel出力（オプション）
openpyxl>=3.1.0
